Finished grocery list UI and searching. Created grocery list, grocery list item, and grocery list view.

+++ lib/models/food.dart
@@ -7,7 +7,7 @@ class Food {
   String? imgUrl;
   String? category;
   String? desc;
-  Float? weight;
+  double? weight;
   bool? ownUnit;
   String? barcode;
